fix(#19): 리뷰 피드백 반영 - V2 구조 통합
- V2로 분리되어 있던 서비스와 리포지토리를 기존 구조로 통합
- 명확한 분리 목적이 없어 불필요한 구조 복잡성 제거
- V2의 개선사항(AvailabilityResult, BundleAvailabilityResult, 도메인 모델 기반 리포지토리)은 기존 구조에 통합
- 테스트 코드 업데이트 및 모든 테스트 통과 확인
